# backend/learning-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Course",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/courses": {
            "get": {
                "tags": [
                    "CourseController"
                ],
                "summary": "Mostrar cursos",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los cursos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "CourseController"
                ],
                "summary": "Subir un curso",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Curso de Programacion"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "01:30:00"
                                    },
                                    "start_date": {
                                        "type": "date",
                                        "example": "2016-01-23 11:53:20"
                                    },
                                    "instructor_id": {
                                        "type": "string",
                                        "example": "61d2082c371e5949897a8fe5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subir un curso a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "put": {
                "tags": [
                    "CourseController"
                ],
                "summary": "Actualizar un curso",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del curso a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "124124124msas1231m"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Curso de Programacion"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "01:30:00"
                                    },
                                    "start_date": {
                                        "type": "date",
                                        "example": "2016-01-23 11:53:20"
                                    },
                                    "instructor_id": {
                                        "type": "string",
                                        "example": "61d2082c371e5949897a8fe5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar un curso a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "CourseController"
                ],
                "summary": "Borrar un curso",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del curso a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "124124124msas1231m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un curso a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/instructors": {
            "get": {
                "tags": [
                    "InstructorController"
                ],
                "summary": "Mostrar instructores",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los instructores."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "InstructorController"
                ],
                "summary": "Subir un instructor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jose"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subir un instructor a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/instructor/{id}": {
            "put": {
                "tags": [
                    "InstructorController"
                ],
                "summary": "Actualizar un instructor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del instructor a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "24214432423v123"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pedro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar un instructor a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/instructors/{id}": {
            "delete": {
                "tags": [
                    "InstructorController"
                ],
                "summary": "Borrar un instructor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del instructor a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "124124124msas1231m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un instructor a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/students": {
            "get": {
                "tags": [
                    "StudentController"
                ],
                "summary": "Mostrar estudiantes",
                "responses": {
                    "200": {
                        "description": "Muestra todos los estudiantes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "StudentController"
                ],
                "summary": "Subir un estudiante",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jose"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subir un estudiante a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/students/{id}": {
            "put": {
                "tags": [
                    "StudentController"
                ],
                "summary": "Actualizar un estudiante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del estudiante a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "24214432423v123"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pedro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar un curso a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "StudentController"
                ],
                "summary": "Borrar un estudiante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del estudiante a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "124124124msas1231m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un estudiante a la base de datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "CourseController",
            "description": "API Endpoints of Courses"
        },
        {
            "name": "InstructorController",
            "description": "API Endpoints of Instructors"
        },
        {
            "name": "StudentController",
            "description": "API Endpoints of Students"
        }
    ],
    "security": [
        []
    ]
}